front: Only delete old extracted ROMs if they exist.

+++ app/src/main/java/paulscode/android/mupen64plusae/GalleryActivity.java
@@ -958,7 +958,7 @@ public class GalleryActivity extends AppCompatActivity implements GameSidebarAct
                         {
                             final File deleteFile = new File(romPath);
 
-                            if(!deleteFile.isDirectory())
+                            if(!deleteFile.isDirectory() && deleteFile.exists())
                             {
                                 deleteFile.delete();
                             }
